head_bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
cffi==1.17.1
discord-protos==0.0.2
discord.py-self==2.0.1
ffmpeg==1.4
frozenlist==1.7.0
idna==3.10
MouseInfo==0.1.3
multidict==6.6.4
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
redis==6.4.0
six==1.17.0
xlib==0.21
xvfbwrapper==0.2.13
yarl==1.20.1
youtube-dl==2021.12.17
